remove bugs vol1
